Move constants package into common package
